passage du prix en float

+++ src/server/Annonce.java
@@ -5,17 +5,17 @@ public class Annonce {
 	private Domain dom;
 	private String titre;
 	private String descriptif;
-	private int prix;
+	private float prix;
 	private int id;
-	
-	public Annonce (Domain d, String titre, String descriptif, int prix, int id) {
+
+	public Annonce (Domain d, String titre, String descriptif, float prix, int id) {
 		this.id = id;
 		this.setDomain(d);
 		this.setTitre(titre);
 		this.setDescriptif(descriptif);
 		this.setPrix(prix);
 	}
-	
+
 	public Domain getDomain() {
 		return this.dom;
 	}
@@ -27,11 +27,11 @@ public class Annonce {
 	public String getDescriptif() {
 		return this.descriptif;
 	}
-	
-	public int getPrix() {
+
+	public float getPrix() {
 		return this.prix;
 	}
-	
+
 	public int getId() {
 		return this.id;
 	}
@@ -39,16 +39,16 @@ public class Annonce {
 	public void setDomain(Domain d) {
 		this.dom = d;
 	}
-	
+
 	public void setTitre(String titre) {
 		this.titre = titre;
 	}
-	
+
 	public void setDescriptif(String descriptif) {
 		this.descriptif = descriptif;
 	}
 
-	public void setPrix(int prix) {
+	public void setPrix(float prix) {
 		this.prix = prix;
 	}
